facemask_detector/requirements.txt
torchvision==0.7.0
matplotlib==3.3.1
pkbar==0.5
torch==1.6.0
beautifulsoup4==4.9.3
Pillow==8.0.0
